Relocate com.pinterest.ktlint.core.RuleExecutionException to com.pinterest.ktlint.core.api.KtLintRuleExecutionException
