+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/api/KtLintRuleExecutionException.kt
@@ -1,12 +1,13 @@
-package com.pinterest.ktlint.core
+package com.pinterest.ktlint.core.api
 
 /**
+ * [KtLintRuleExecutionException] is thrown whenever an error occurs during execution of a KtLint [Rule].
+ *
  * @param line line number (one-based)
  * @param col column number (one-based)
  * @param ruleId rule id (prepended with "&lt;ruleSetId&gt;:" in case of non-standard ruleset)
  */
-@Deprecated("Marked for removal from public API in KtLint 0.48. Please raise an issue if you have a use case to keep it public.")
-public class RuleExecutionException(
+public class KtLintRuleExecutionException(
     public val line: Int,
     public val col: Int,
     public val ruleId: String,
